Retire `createConvertMemRefToTppPass`
It is better to convert memref to xsmm directly without passing via TPP
operations. This simplifies the pipeline (see `tensor-to-unary.mlir`)
and avoid TPP dependency.
