+++ include/TPP/Dialect/Xsmm/XsmmUtils.h
@@ -14,11 +14,28 @@
 namespace mlir {
 class Type;
 class RewriterBase;
+class Value;
+class ArrayAttr;
+class Operation;
 
 namespace xsmm {
+class UnaryKindAttr;
+
+struct UnaryInfo {
+  unsigned m;
+  unsigned n;
+
+  int64_t ldi;
+  int64_t ldo;
+};
+
 namespace utils {
 
-xsmm::DataTypeAttr getDataType(RewriterBase &rewriter, Type type);
+DataTypeAttr getDataType(RewriterBase &rewriter, Type type);
+FailureOr<UnaryInfo> getUnaryInfo(Value input, Value output);
+void replaceOpWithUnary(RewriterBase &rewriter, Operation *operation,
+                        ArrayRef<Value> operands, UnaryInfo unaryInfo,
+                        ArrayAttr flags, UnaryKindAttr kind);
 
 } // namespace utils
 } // namespace xsmm
