+++ include/TPP/Passes.h
@@ -148,7 +148,7 @@ createGpuToCudaPass(StringRef gpuTriple = "nvptx64-nvidia-cuda",
 std::unique_ptr<OperationPass<ModuleOp>> createGpuToVulkanPass();
 std::unique_ptr<OperationPass<func::FuncOp>>
 createRewriteBatchMatmulToMatmulPass();
-std::unique_ptr<OperationPass<func::FuncOp>> createConvertMemRefToTppPass();
+std::unique_ptr<OperationPass<func::FuncOp>> createConvertMemRefToXsmmPass();
 std::unique_ptr<OperationPass<ModuleOp>>
 createDefaultPipelinePass(StringRef gpuBackend = "");
 std::unique_ptr<OperationPass<ModuleOp>>
