+++ lib/TPP/Dialect/Xsmm/XsmmUtils.cpp
@@ -7,6 +7,8 @@
 //===----------------------------------------------------------------------===//
 
 #include "TPP/Dialect/Xsmm/XsmmUtils.h"
+#include "TPP/Dialect/Xsmm/XsmmOps.h"
+#include "TPP/ValueUtils.h"
 #include "mlir/IR/PatternMatch.h"
 #include "mlir/IR/TypeUtilities.h"
 
@@ -14,12 +16,58 @@ namespace mlir {
 namespace xsmm {
 namespace utils {
 
-xsmm::DataTypeAttr getDataType(RewriterBase &rewriter, Type type) {
+DataTypeAttr getDataType(RewriterBase &rewriter, Type type) {
   auto elemType = getElementTypeOrSelf(type);
   if (elemType.isBF16())
     return xsmm::DataTypeAttr::get(rewriter.getContext(), xsmm::DataType::BF16);
   assert(elemType.isF32() && "Element type neither bf16 nor f32");
   return xsmm::DataTypeAttr::get(rewriter.getContext(), xsmm::DataType::F32);
+}
+
+void replaceOpWithUnary(RewriterBase &rewriter, Operation *operation,
+                        ArrayRef<Value> operands, UnaryInfo unaryInfo,
+                        ArrayAttr flags, xsmm::UnaryKindAttr kind) {
+  Location loc = operation->getLoc();
+  IntegerType integer64 = IntegerType::get(rewriter.getContext(), 64);
+  DenseI64ArrayAttr dims = DenseI64ArrayAttr::get(
+      rewriter.getContext(), ArrayRef<int64_t>{unaryInfo.m, unaryInfo.n,
+                                               unaryInfo.ldi, unaryInfo.ldo});
+  auto dtype = xsmm::utils::getDataType(rewriter, operands.back().getType());
+  Value dispatched = rewriter.create<xsmm::UnaryDispatchOp>(
+      loc, integer64, kind, dims, flags, dtype);
+  SmallVector<Value> invokeOperands;
+  invokeOperands.push_back(dispatched);
+  invokeOperands.append(operands.begin(), operands.end());
+  rewriter.replaceOpWithNewOp<xsmm::UnaryOp>(operation, dtype, kind,
+                                             invokeOperands);
+}
+
+FailureOr<UnaryInfo> getUnaryInfo(Value input, Value output) {
+  Type outputType = output.getType();
+
+  assert(isa<ShapedType>(outputType));
+  auto outputShapedType = output.getType().cast<ShapedType>();
+  if (outputShapedType.getRank() != 2)
+    return failure();
+
+  UnaryInfo unaryInfo;
+  unaryInfo.m = outputShapedType.getShape()[0];
+  unaryInfo.n = outputShapedType.getShape()[1];
+
+  int64_t ldi = 1;
+  if (isa<ShapedType>(input.getType())) {
+    auto stridesOnInput = mlir::utils::getStaticStrides(input);
+    if (failed(stridesOnInput) || stridesOnInput->back() != 1)
+      return failure();
+    ldi = stridesOnInput->front();
+  }
+  auto stridesOnOutput = mlir::utils::getStaticStrides(output);
+  if (failed(stridesOnOutput) || stridesOnOutput->back() != 1)
+    return failure();
+
+  unaryInfo.ldi = ldi;
+  unaryInfo.ldo = stridesOnOutput->front();
+  return unaryInfo;
 }
 
 } // namespace utils
